fix: move functions to own file to allow serverside import/usage in next.js app
